added the add and commit commands to gitr and verified that they are compatible with git

+++ lib/git-ruby/raw/internal/loose.rb
@@ -74,11 +74,13 @@ module GitRuby
           sha1 = Digest::SHA1.hexdigest(store)
           path = @directory+'/'+sha1[0...2]+'/'+sha1[2..40]
           
-          content = Zlib::Deflate.deflate(store)
+          if !File.exists?(path)
+            content = Zlib::Deflate.deflate(store)
           
-          FileUtils.mkdir_p(@directory+'/'+sha1[0...2])
-          File.open(path, 'w') do |f|
-            f.puts content
+            FileUtils.mkdir_p(@directory+'/'+sha1[0...2])
+            File.open(path, 'w') do |f|
+              f.write content
+            end
           end
           return sha1
         end
